+++ src/interactions/actions/drill-down.ts
@@ -94,10 +94,16 @@ export class DrillDownAction extends Action {
 
     const { position } = this.getButtonCfg();
 
-    // 默认，左上角直接出发
-    let point = { x: 0, y: 0 };
+    // @todo 后续抽取一个函数来处理，以及增加 margin 或者 padding 的设置
+    // 非 polar 的，需要使用 coordinate，除却图表组件
+    let point = { x: coordinate.start.x, y: coordinate.end.y - (bbox.height + PADDING_TOP * 2) };
+    if (coordinate.isPolar) {
+      // 默认，左上角直接出发
+      point = { x: 0, y: 0 };
+    }
     if (position === 'bottom-left') {
-      point = coordinate.isPolar ? { x: 0, y: coordinate.getHeight() } : coordinate.convert({ x: 0, y: 1 });
+      // 涉及到坐标反转的问题
+      point = { x: coordinate.start.x, y: coordinate.start.y };
     }
     /** PADDING_LEFT, PADDING_TOP 与画布边缘的距离 */
     const matrix = Util.transform(null, [['t', point.x + PADDING_LEFT, point.y + bbox.height + PADDING_TOP]]);
@@ -156,7 +162,7 @@ export class DrillDownAction extends Action {
     }
 
     const { view } = this.context;
-    const data = last(historyCache).children;
+    const data = last(historyCache).children; // 处理后的数组
     view.changeData(data);
 
     if (historyCache.length > 1) {