+++ src/plots/radial-bar/adaptor.ts
@@ -4,7 +4,7 @@ import { flow, deepAssign, findGeometry, transformLabel } from '../../utils';
 import { interval, point } from '../../adaptor/geometries';
 import { processIllegalData } from '../../utils';
 import { RadialBarOptions } from './types';
-import { getScaleMax } from './utils';
+import { getScaleMax, getStackedData } from './utils';
 
 /**
  * geometry 处理
@@ -49,9 +49,10 @@ function geometry(params: Params<RadialBarOptions>): Params<RadialBarOptions> {
  */
 export function meta(params: Params<RadialBarOptions>): Params<RadialBarOptions> {
   const { options } = params;
-  const { yField, maxAngle, data } = options;
+  const { yField, xField, data, isStack, isGroup, colorField, maxAngle } = options;
 
-  const processData = processIllegalData(data, yField);
+  const actualData = isStack && !isGroup && colorField ? getStackedData(data, xField, yField) : data;
+  const processData = processIllegalData(actualData, yField);
   return flow(
     scale({
       [yField]: {