+++ src/utils/hierarchy/util.ts
@@ -1,4 +1,12 @@
-import { isArray, isString } from '@antv/util';
+import { isArray, isString, filter } from '@antv/util';
+
+/** export 一些字段常量 */
+/** 在同层级，同一父节点下的节点索引顺序 */
+export const NODE_INDEX_FIELD = 'nodeIndex';
+/** child 节点数量 */
+export const CHILD_NODE_COUNT = 'childNodeCount';
+/** 节点的祖先节点 */
+export const NODE_ANCESTORS_FIELD = 'nodeAncestor';
 
 const INVALID_FIELD_ERR_MSG = 'Invalid field: it must be a string!';
 
@@ -32,8 +40,25 @@ export function getField(options: Options, defaultField?: string): string {
 export function getAllNodes(root: any) {
   const nodes: any[] = [];
   if (root && root.each) {
-    // d3-hierarchy
+    let parent;
+    let index;
+    // d3-hierarchy: Invokes the specified function for node and each descendant in **breadth-first order**
     root.each((node: any) => {
+      if (node.parent !== parent) {
+        parent = node.parent;
+        index = 0;
+      } else {
+        index += 1;
+      }
+      const ancestors = filter(
+        (node.ancestors?.() || []).map((d: any) => nodes.find((n) => n.name === d.name) || d),
+        ({ depth }) => depth > 0 && depth < node.depth
+      );
+
+      node[NODE_ANCESTORS_FIELD] = ancestors;
+      node[CHILD_NODE_COUNT] = node.children?.length || 0;
+      node[NODE_INDEX_FIELD] = index;
+
       nodes.push(node);
     });
   } else if (root && root.eachNode) {